Strategic typing improvements: targeted error reduction
Applied strategic typing framework achieving additional 4-error reduction
(287 → 283 errors) through targeted return type fixes and cleanup.

Key improvements:
- Strategic float() casting for NumPy return types in domain_3d geometry
- Cleaned up unused type: ignore comments in domain_2d and jax_utils
- Fixed floating[Any] → float return type compatibility
- Removed obsolete import annotations after stub creation

Technical patterns applied:
- Strategic casting: float(max(edges) / 2.0) for NumPy compatibility
- Cleanup methodology: Remove type: ignore after stub infrastructure
- Return type compatibility: Consistent float returns in geometry calculations
- Import cleanup: Leverage enhanced stub infrastructure

Files improved: 3 core modules with geometry and acceleration components
Impact: Better type safety in numerical computations + cleaner codebase

+++ mfg_pde/geometry/domain_3d.py
@@ -471,7 +471,7 @@ class Domain3D(BaseGeometry):
                 edge_length = np.linalg.norm(coords[i] - coords[j])
                 edges.append(edge_length)
 
-        return max(edges) / 2.0  # Approximation
+        return float(max(edges) / 2.0)  # Approximation
 
     def _compute_inradius(self, coords: np.ndarray) -> float:
         """Compute inradius of tetrahedron."""
@@ -495,7 +495,7 @@ class Domain3D(BaseGeometry):
             edge1 = face[1] - face[0]
             edge2 = face[2] - face[0]
             area = 0.5 * np.linalg.norm(np.cross(edge1, edge2))
-            surface_area += area
+            surface_area += float(area)
 
         # Inradius = 3 * Volume / Surface_area
         if surface_area > 0:
